Description
Pick n Pay vending system for daily operations. This project will explain more in details on how the this company sell their products , keep track of stock, and receive some prompts from the system to reoder some new item if the stock runs low. It show how the manager be able to add new stock before selling begins. Is anable to manage employees, particularly on the hours worked, which can be used to calculate their salaries depending on the job grade of the employee. It track whether an employee is permanently employed or is on contract.